Add context menu for installed_list_widget.py
